Changed: internal organization in IupLua for better nomenclature.
git-svn-id: svn://svn.code.sf.net/p/iup/iup/trunk@3016 2f08bc20-984a-4df8-b195-9b5d5d477537

+++ iup/srclua5/elem/il_image.c
@@ -65,8 +65,8 @@ static int Image (lua_State * L)
   }
   else
   {
-    w = luaL_checkinteger(L, 1);
-    h = luaL_checkinteger(L, 2);
+    w = (int)luaL_checkinteger(L, 1);
+    h = (int)luaL_checkinteger(L, 2);
     pixels = iuplua_checkuchar_array(L, 3, w*h);
     ih = IupImage(w, h, pixels);
     free(pixels);
